+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changes
+- V4.0.5
+	- WARNING: You may want to backup your config prior to launching this version of the app as config settings have changed.
+	- Changed configuration settings to use bool instead of string for binary configuration settings [On/Off] 
 - V4.0.4
 	- Added terminal command option. (Thanks to [UmaLPZ](https://github.com/UmaLPZ) for the suggestion)
 	- See [UseOnlyTerminal] configuration setting to disable keybinds and use only the terminal for cleanup